node: force host-deploy to run over a custom ssh port
Force host-deploy to run over a custom ssh port at
registration time.
With this we can map a custom host port to SSH on
the node container and have the engine run
host-deploy over the host address reaching our
container getting the SSL cert signed for
the host address so that the communication to
the vdsm port over the host address will
no fail due to SSL verification.

Change-Id: I5e1079efe07cb8620992848132697c02fa154e15
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>