engine: fixing pki handling for successive executions
Ensuring the engine container can run a second time with the
PKI created by the first run of engine-setup:
on the first run engine-setup creates the oVirt PKI writing it
on a persistent volume. engine-setup requires a template there
but the initial persistent volume is by default empy and so we
have to copy the template there on the first execution and avoid
overwriting it on successive ones.
We have also to force the ownership of the PKI volume since
in kubernetes the ownership of mounted volumes is always
set at root while the engine requires ovirt:ovirt. See
https://github.com/kubernetes/kubernetes/issues/2630

Change-Id: Ic154f10c5c9b53846f9b964879c6c5d59c57c5c1
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>