Use 'IfNotPresent' image pull policy for engine
Currently the image pull policy for engine containers is 'Always'. This
complicates testing, as it isn't possible to build the image locally and import
it manually to the docker instance used for tests. For example, when using
minishift one can rebuild and test the engine image as follows:

  $ eval $(minishift docker-env)
  $ cd image-specifications/engine
  $ docker build -t docker.io/bronhaim/ovirt-engine:4.1 .

That will build the new image using the docker daemon tht runs inside the
minishift virtual machine, so the image will be already there, no need to pull
it from any registry. But when deploying the application OpenShift will anynow
try to pull it, and then it will use the old one.

To avoid this issue This patch changes the policy to 'IfNotPresent'.

Change-Id: Ic45fecf8298f248c1bd299cd74fe9c5e4c26682c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>