Don't install the OVN provider
The installation of the OVN provider has been recently added to the
engine setup tool. As this tool runs during the start of the container
and there is no answer in the 'answers.conf' file, the container fails
to start. To avoid that problem this patch modifies the 'answers.conf'
file so that the OVN provider isn't installed.

Change-Id: If89084a1d67c40b5a44d728881a2c434deb187f4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>