Add 'autovacuum' parameters to the PostgreSQL image
Recent versions of the engine need to have the following auto-vacuum
configuration parameters:

  autovacuum_vacuum_scale_factor = 0.01
  autovacuum_analyze_scale_factor = 0.075
  autovacuum_max_workers = 6
  maintenance_work_mem = 65536

Without them the setup tool fails during the start of the engine container.

This patch modifies the 'ovirt-db-config.sh' file so that those parameter are
automatically added to the 'postgresql.conf' file during the start of the
PostgreSQL container.

Change-Id: I90924a3326d478611f7f2962e2a641968da86023
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>