Don't perform database full vacuum
The engine setup tool has been recently modified to ask the user if a full
vacuum of the database should be performed. As there is no answer for this in
the 'answers.conf' file, the engine start fails. This patch modifies the
answers file so that the full vacuum isn't performed.

Change-Id: I3b7099ef609831660d7602a18f715fa85426f0f4
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit c213874adc4a51cfce481bba469a3ac43baea128)