Add 'base' image
This patch adds a new 'base' image intended to be used by all the oVirt
containers that need to install RPM packages distributed by oVirt. This
image is based on CentOS 7, and it just installs the oVirt release RPM
and other packages that are needed by all oVirt containers (currently
none).

Change-Id: I792cbba4b5b698b594d768964a01b016e3372f01
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>