Add build tool
This patch adds a build tool that loads the source files of the images,
checks their dependencies, and builds them in the right order. The build
tool is witten in Go, and lives in the 'tools' directory.

To simplify the build and execution of the build tool, the patch also
adds a Makefile, so building the project is a matter running the 'make'
command. That takes care of building the build tool, if needed, and then
building the images.

The tags assigned to the images are configurable using the 'build.conf'
file. For example to build images tagged 'my/engine:4.1' instead of the
default 'ovirt/engine:master', edit the 'build.conf' file so that it
contains this:

  prefix=my
  version=4.1

The run the 'make' command.

Change-Id: Ic0f8e0374264f43b5e7a23c22e5c134ce4e5d3a6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>