Use the 'base' image as the parent of the SPICE proxy
This patch changes the SPICE proxy image so that it uses the 'base'
image as its parent. That is an step in the direction of using the same
base image for all the containers in the project.

Change-Id: I19b26ae3c520ce722184769023958fc5d11fcfb8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 5e7ca0c8129439fcc27c53b9f55c59671b6a4e26)