Add support for a private Docker registry
This patch adds a new 'registry' parameter to the 'build.conf' file.
That parameter can be used to set the address of the Docker registry
where the 'push' tool will push the images:

  registry=localhost:5000

The default will be 'localhost:5000'. That registry can be created
easily with the following command:

  docker run -d -p 5000:5000 --restart=always --name registry registry:2

Refer to the Docker registry documentation for details:

  https://docs.docker.com/registry/deploying

Change-Id: Ia4d80bc51ac53b3fddb72d53f22c6e734ec6c7b5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>