Fail automatition script if build fails
Currently the automation scripts return the exit code of the cleaning
commands. That means that builds that should have been considered as
failures are considered as sucesses. To avoid that this patch changes
the automation scripts so that they return the exit code of the last
build step *before* cleaning.

Change-Id: Idedbe1c92865579bfd68995f486d54dcd40b53f6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>