Use softwarecollections.org PostgreSQL image
This patch changes the engine database container to use the
softwarecollections.org PostgreSQL 9.5 image. See here for details:

  https://www.softwarecollections.org/en/scls/rhscl/rh-postgresql95
  https://github.com/sclorg/postgresql-container

This simplifies using other versions of the PostgreSQL image based on
softwarecollections.org. For example, to use the image based on RHEL 7
provided by Red Hat the only required change is to use this as the base:

   registry.access.redhat.com/rhscl/postgresql-95-rhel7

Change-Id: I2c4e135aa4628391b73621115996fd3ad75d9629
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>