Add 'deploy' tool
This patch adds a simple deploy tool that basically runs the deployment
steps described in the README file. Currently it doesn't do much more
than that, and it doesn't provide much benefit, but it is the seed for
what should become the tool used to deploy the application in
development and production environments. To use it just run the
following:

  make deploy

Change-Id: Ib9a13bed3cfb9ed875ce3e6df649f3c6963b50e1
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>