Add 'clean' tool
Currently the 'clean' target of the build process is implemented inside
the Makefile, and it removes the dangling images. But it doesn't remove
the images that have been created by the 'build' target. That means that
doing something like this:

  make clean && make build

Will not really build the images from scratch, as the intermediate steps
will be still saved in the local docker storage. In order to make that
work this patch introduces a new 'clean' tool wich explicitly removes
all the images. This will be called from the Makefile, right before
removing the dangling images.

Change-Id: I2406e9b81bfadd79d9332cc1e7987177f1a4da2a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>