Introduce 'project' concept
Currently the configuration needed to build the project is
stored in a 'build.conf' file. Some of the settings in that
file are or will be not related to building, but to other
tasks. For example, in the future we will introduce
parameters that are related to the deployment of the
containers, not to the build of the images. In order to
better support that this patch renames the file to
'project.conf' and introduces a new 'Project' type that will
contain the information loaded from that file, from the
image specifications, and from the OpenShift manifests.
Instances of this type will be initially loaded and passed
to the different tools that need them.

Change-Id: I90d63422771c05e67f0b6efb3fbb803187e4cf15
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>