Unify tools
Currently all the tools (build, deploy, push, etc) have the same
structure: they first load the project configuration and then they
perform some tasks. In addition a different binary is generated
for each tool, and compiled only when the tool is used. That means that
the code that loads the project configuration is repeated, and that the
tools are compiled/checked only when used. In order to improve that this
patch introduces a new 'tool' binary that receives as parameter the name
of the tool. This main binary takes care of loading the project
configuration and running the specific tool.

Change-Id: I3eba76bca19b89bb73d0816889f9bddbde11af6a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>