Use Glide to manage Go dependencies
Currently the Go dependencies of the project are running 'go get',
from the Makefile. This doesn't scale well for adding other
dependencies. To simplify that this patch changes the project so that it
uses Glide to manage dependencies:

  https://glide.sh

If Glide isn't installed it will be installed automatically when running
the build.

Change-Id: Iae3eb9ddd399502b7afdbd8e2f5ed5ecadca335e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>