Improve console and log output
This patch changes the tools so that they automatically generate a log
file that contains all the details of the execution. The console will
now only contain the more relevant messages. For example, the 'build'
tool will only present the following messages in the console:

  [INFO] Log file is '/projects/ovirt-containers/build.log'
  [INFO] Loading project file '/projects/ovirt-containers/project.conf'
  [INFO] Building image 'ovirt/base:master'
  [INFO] Building image 'ovirt/engine:master'
  [INFO] Building image 'ovirt/engine-database:master'
  [INFO] Building image 'ovirt/engine-spice-proxy:master'
  [INFO] Building image 'ovirt/vdsc:master'
  [INFO] Building image 'ovirt/vdsc-syslog:master'
  [INFO] Tool finished successfully

The rest of the details, including the commands executed and their
output, will be available in the 'build.log' file.

In addition the messages presented in the console will have a touch of
color. The [INFO] prefix will be green and the [ERROR] prefix will be
red.

Change-Id: I85a6887693df9771cc5f3810db62d6371e34b121
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>