Add template support
Currently the Dockerfiles and the OpenShift manifests that are part of
the project contain hardcoded references to images. For example, the
Dockerfile for the engine contains the following reference to the 'base'
image:

  FROM ovirt/base:master

And the OpenShift manifest for the engine contains the following
reference to the 'engine' image:

  image: ovirt/engine:master

That complicates usage of the images and the manifests in environments
where it is necessary or convienient to use different registries,
different prefixes or different versions. For example, lets say that we
need to change the 'project.conf' configuration file like this:

  version=4.2.1

  [images]
  prefix=my
  registry=myregistry.example.com:5000

The only way to have correct image dependencies and correct OpenShift
manifests is then to manually (or with a script) change the source
files.

To simplify that this patch introduces support for generating all the
files of the project from templates, using the template syntax
implemented in the Go templates package:

  https://golang.org/pkg/text/template

Files don't need to have any special extension or name to be considered
templates: all of them will be processed as templates.

Templates will have available a context containig a 'tag' function that
can be used to get the tag corresponding to certain image. That means
tha the Dockerfile for the engine can be written as follows:

  FROM {{ tag "base" }}

That will be translated taking into account the content of the
'project.conf' file:

  FROM myregistry.example.com:5000/my/engine:4.2.1

Same for OpenShift manifests. The engine manifest can then be like this:

  image: {{ tag "engine" }}

And it will be translated into this:

  image: myregistry.example.com:5000/my/engine:4.2.1

An important consequence of this is that OpenShift manifest files won't
be usable directly, only via the the 'make deploy' mechanism.

Change-Id: I394f896441cf9fcc6792de5e12b708451cfdc379
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>