Embed project configuration
Currently the tools that build or deploy the project need to
be executed in a directory that contains the 'project.conf'
file, as well as the rest of the source files of the
project. That complicates things for users that just want to
download and test the containers, as they need to download
the complete set of files. In order to simplify that this
patch changes the tool so that all the required files are
embedded within the binary of the tool. The tool will now
check if the 'project.conf' file is available. If it isn't
it will extract the embedded files to a temporary directory
and use them. This means that the user will only need to
download the 'ovc' binary and then run 'ovc deploy'.

Change-Id: I6fec0b5aaa432030f9ce6ef142a06f0ab8ec781d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>