Use relative paths
Currently the objects that contain the information about the
project have the 'directory' and 'work' field, containing
the absolute paths of the source and work directories for
those objects. Using absolute directories complicates moving
the project to a different directory, which is needed by the
next patch adding support for embedding the project
configuration inside the binary. In order to address that
this patch replaces 'directory' with 'root' (absolute, only
for the project) and 'path' (relative, for everything else).
The 'work' attribute is removed, and instead work
directories are calculted as needed.

Change-Id: I9005bec9791125348b0f03f7f8a0c3edeea2953c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>