Use 'defer' mechanism to close log and project
Currently the main program closes the log files and the
project explicitly. That isn't a problem, but it doesn't
scale well if additional things need to be released. In
order to improve that this patch changes the main program
so that it uses the 'defer' mechanism, which will be easier
to use in later patches that will need to release other
resources.

Change-Id: I033bd3440309de25d6a4ffaad53787a744adfec6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>