Don't require specific 'epel-release' version
Currently the VDSC image downloads the 'epel-release'
package specifying directly the URL. That URL has recently
started to fail, because the file has been replaced by a
newer version. To avoid that problem now and in the future,
this patch changes the VDSC image so that it just uses the
normal yum mechanism to to find, download and install the
latest version of the package.

Change-Id: If1a3796ff5ecd7124aed438fc95c91b3e225c68f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>